Move a couple of functions
